+++ src/ctf01d/employees/employ_web_server.h
@@ -63,6 +63,7 @@ private:
   int httpApiV1Scoreboard(HttpRequest* req, HttpResponse* resp);
   int httpApiV1GetPaths(HttpRequest* req, HttpResponse* resp);
   int httpApiV1Flag(HttpRequest* req, HttpResponse* resp);
+  int httpApiV1Metrics(HttpRequest* req, HttpResponse* resp);
   int httpLogo(const std::string &request_path, HttpRequest* req, HttpResponse* resp);
 
   std::shared_ptr<hv::HttpService> m_pHttpService;